Refine code:
1. Use stb to load texture.
3. Use uint8 to store texture.
2. Calculate the FPS.

+++ tr.h
@@ -9,7 +9,7 @@
 
 struct TRTexture
 {
-    float *data;
+    uint8_t *data;
     int stride;
     int w;
     int h;
